+++ modules/linear_boltzmann_solvers/discrete_ordinates_problem/sweep/angle_set/aahd_angle_set.cu
@@ -12,7 +12,7 @@ namespace opensn
 {
 
 AAHD_AngleSet::AAHD_AngleSet(size_t id,
-                             size_t num_groups,
+                             unsigned int num_groups,
                              const SPDS& spds,
                              std::shared_ptr<FLUDS>& fluds,
                              std::vector<size_t>& angle_indices,
