+++ modules/linear_boltzmann_solvers/discrete_ordinates_problem/sweep/fluds/fluds.h
@@ -20,7 +20,7 @@ class SPDS;
 class FLUDS
 {
 public:
-  FLUDS(size_t num_groups, size_t num_angles, const SPDS& spds)
+  FLUDS(unsigned int num_groups, size_t num_angles, const SPDS& spds)
     : num_groups_(num_groups),
       num_angles_(num_angles),
       num_groups_and_angles_(num_groups_ * num_angles_),
@@ -60,7 +60,7 @@ public:
   virtual ~FLUDS() = default;
 
 protected:
-  const size_t num_groups_;
+  const unsigned int num_groups_;
   const size_t num_angles_;
   const size_t num_groups_and_angles_;
   const SPDS& spds_;
